image.py: major changes to cropping/autocrop
major breaking changes to cropping, including renaming some
functions for clearer indication of their purpose.

The original `crop` function has been rename to `centred_crop`
to indicate better what it does. It no longer has any knowledge
of downsampling, and just forces the cropped region to be the size
that was asked for, at the expense of it being centred on the
requested centre.

`crop` is now a very simple function that crops to limits given.

`autocrop` will now not raise RuntimeError if optimal crop is not
found; instead, use the best effort guess.