ui: remove references to fit region
Now that autocrop is being used, fit region does not need to be
changed manually and so widgets can be removed. Kept this removal
in a separate commit in case we decide to add it back. However, in
the case where autocrop is not providing an image that can be
fitted, it probably makes more sense to tune the parameters of the
autocrop instead