image: parameter dictionary name change
x0 is an unfortunate designation for the centroid. Since this will
only be used with images, use a more sensible name: pxc, aka
PiXel Centre. It's not entirely obvious, but at least it's not
actually misleading - is x0 the x position of the centre of the
beam?

While changing the name of this, it also made sense to change some
variable names in image.py to make the code a little clearer.

Also removed default args for parameters - they don't make all
that much sense anyway, may as well raise an error if someone
hasn't provided all the required parameters.