image: auto_crop accepts fill_target argument
The fill_target argument affects how much of the image autocrop
will attempt to make the beam fill. This can make fits more/less
robust, the figure of 0.3 was chosen fairly arbitrarily, mostly
because it looks good.

Might be worth optimising this choice, but for now we can just
leave it up to the user if they want.