image: fit optimizations
Minor optimizations; the first uses a fixed path for the
einstein summation (unfortunately not a great win, but a win
nonetheless).

The second optimization regards the `scale` parameter. Originally
this was intended as a sort of `max beam intensity` parameter,
as can be seen in the initial guess for it
(array max - array min). However, in the fitting function I had
also divided by the normalization for unit integral over all
space of the Gaussian, making it into an integrated power.
This meant that the initial guess was always a wild underestimate,
which often caused the fit to fail due to maximum number of
function evaluations. Removing this scaling prefactor improves
the robustness of the fitting.